third_party/blink/web_tests/virtual/boost-image-priority/http/tests/priority-hints/image-initial-load-expected.txt

This is a testharness.js-based test.
PASS all images must be fetched by the preload scanner
PASS high fetchpriority on <img> must translate to kHigh resource load priority
PASS low fetchpriority on <img> must translate to kLow resource load priority
FAIL auto fetchpriority on early <img> must translate to kMedium resource load priority assert_equals: expected 2 but got 1
FAIL invalid fetchpriority on early <img> must translate to kMedium resource load priority assert_equals: expected 2 but got 1
FAIL missing fetchpriority on early <img> must translate to kMedium resource load priority assert_equals: expected 2 but got 1
Harness: the test ran to completion.
